move non-public classes into essentials.py
